feat: added default policy and forward default selector to handle selection of multiple schemas
Although this HC solution works it does not match the MVC controller in terms of functionality. A requirement cannot just be a composition of a schema, and required authentication. How the HC solution works is by adding an additional requirement to the JWT claims. For the bearer two token, or policy `JustBearerTwo` we check if the role is standard. If the requirement is not met, the request is invalid, meaning `BearerOne` token will not succeed because it's missing role = standard.